made singleton

+++ geometry/TrianglesView.swift
@@ -6,14 +6,14 @@
 
 import UIKit
 protocol TrianglesViewDataSource: class {
-    func getColor(sender: TrianglesView) -> [CGFloat]?
+    func getColor(sender: TrianglesView) -> [Float]?
 }
 
 @IBDesignable
 class TrianglesView: UIView  {
     
     @IBInspectable
-    var colorRGB = [CGFloat]() { didSet { setNeedsDisplay() } }
+    var colorRGB = [Float]() { didSet { setNeedsDisplay() } }
         
     private var dataSource: TrianglesProtocol?
     
@@ -24,7 +24,11 @@ class TrianglesView: UIView  {
         CGContextSetLineWidth(context, 3)
         if let color = dataSourceColor?.getColor(self) {
             colorRGB = color
-            CGContextSetRGBStrokeColor(context, colorRGB[0], colorRGB[1], colorRGB[2], 1.0)
+            CGContextSetRGBStrokeColor(context,
+                                       CGFloat(colorRGB[0]),
+                                       CGFloat(colorRGB[1]),
+                                       CGFloat(colorRGB[2]),
+                                       1.0)
         } else {
             UIColor.greenColor().set()
         }
